// job/src/types/WorkerState.ts
export enum WorkerState {
  INITIALIZING = "INITIALIZING",
  CONNECTING = "CONNECTING",
  ACTIVE = "ACTIVE",
  RECONNECTING = "RECONNECTING",
  PHONE_VERIFICATION = "PHONE_VERIFICATION",
  EXPECTING_CODE = "EXPECTING_CODE",
  ERROR = "ERROR",
  SHUTDOWN = "SHUTDOWN",
  BANNED = "BANNED",
  CONNECTION_LOST = "CONNECTION_LOST",
}